OpenCore: Enable PV registry to distinguish AMR-NB and AMR-WB decoder components
Currently in PV OMX registry AMR component is common for both AMR-NB and
WB decoders. Ideally there should be 2 individual componets for NB and
WB, to ensure that if NB is hardware accelerated, WB can be Software and
vice versa. So registering as two individual components to get that
flexibility.