OpenCore: Fix for video playback pause/resume delay for a long clip
For a long clip, when we try to resume from the point video playback was left off, actualMediaDataTS does not get updated.  As a result all the intermediate frames between this time and sync point are decoded.  This introduces a delay in video playback.
This fix updates the actualMediaDataTS to the sync point, decode begins from the sync point avoiding delay in video playback.

CRs-Fixed: 182892