Opencore: CameraMIO: Avoid a rare deadlock condition while Stopping Camcorder Record
This deadlock happens only when the application has issued a DoStop and
the camera driver has sent the recorded frame, before the DoStop was
completely processed. In this case, we should not release the recording
frame in the same thread context which will lead to a deadlock. The
frame will be released during stop, which has already been issued.

CRs-Fixed : 186543