external/opencore: fix the private list initialiation and its passing to component
One private list, st_list, is being used for all the input buffers,
which is leading to data corruption, when input buffers are queued at
the component. To fix this, allocate as many private lists as equal to
input buffers and establish one-to-one mapping between private lists and input buffers.

CRs-Fixed: 189835