OpenCore: Setting of cropping window from container file
For h263, there are some clips that contain width/height information in
container files which is different from the byte stream. This information is
available in source node which parses the container tracks information, however
it was not passed to decoder node. Enabled passing of these parameters to
decoder node and if a display width/height is passed, use it as a cropping
window, getting rid of possible garbage displayed on screen.

CRs-Fixed: 184452