OpenCore: Fix first frame dropped for mpeg4 video recording
Hardware encoder sends VOL header together with first I-Frame for the first
output frame. Opencore expects the first output buffer from encoder to contain
just the VOL header and then dimisses the rest of the buffer, causing the first
frame to be dropped from clip.

Changed the way Opencore parses the first output buffer, reading the VOL header
and then pass the rest of the data (if any) to composer.

CR-s Fixed: 188705,189870