opencore: Changes made to allow the last audio packet to write even after STOP command is issued by player
In the current PV meadia node there is only check for state .Added one
more check for data len which will help to write the last packet and
will not enter oscl_leave() which causes PV to abort and crash. This
change fixes the issue of 10 mins recorded clip being bad and unable to
play and there is crash and no proper close of PV player.

CRs-Fixed : 190123