Opencore: Correcting android surface output to support software codecs correctly.
The frame buffer index was not being set correctly at the beginning
before posting the buffer to SurfaceFlinger. Also when trying to refresh
the last frame with pausing, buffer index offset for hardware was being
used instead of software buffer index offset. This cause the video to
look strange when starting a clip and pausing.

CRs-Fixed: 212041