Opencore: Using Movable flag to avoid corruption and jerkiness in recorded clip for video.
Making use of iOMXComponentSupportsMovableInputBuffers allows encoder
node to increment the refrence count for the buffer/frag when sending it
to OMX component which can be reduced only when Empty buffer done is
called for that particular buffer and buffer is released to camera by
unbind mechanism.
This handshaking avoid the overwriting of input buffers by camera which
protect the recorded clip from corrupting in USebuffer case.

CRs-Fixed: 214246,191046