Opencore: Ignoring frame duration if timestamps are not in order
Duration was not being calculated correctly in the parser in cases where
the clip contained B or P frames. The timestamps in these cases were not
in order, and the miscalculated duration would cause issues where frames
would drop. This change will make the parser ignore the duration by
setting it to 0 only if the timestamps are not in order.

Change-Id: I91a4bd58193a854cef517e156e930f31af5705ab
CRs-Fixed: 228020