Incorrect timestamps in camera samples leads to record stop
AudioInput MIO updates authorclock with system time when the
first audio sample was received. This is used as an offset by
CameraInput MIO to calculate timestamps for camera samples.

Currently the code to calculate the system time in AudioInput
MIO is executed if the format type is PCM_16_BIT only.
A 0 value is assumed for other formats. This can lead to
incorrect timestamps for camera samples. During a record, the composer node
checks the timestamp of each camera sample with a maximum duration value.
If this check fails, the author engine issues a stop in the middle
of the record.

This fix makes calculation of the system time in AudioInput MIO common
to all formats thus preventing incorrect timestamps for camera samples.

CRs-Fixed: 235680

Change-Id: If424cfa81bb6bbe27064b657591977d130a981a3