OpenCore: playback failure for non-PBCtrl nodes
Initial return code in this piece of code is set to 'Failure'.
Then, return code is modified according to the result of the
NotifyTargetPositionSync() function.
However, this function is only called for nodes that support
repositioning. For nodes that don't, the initial return code
will stay 'Failure', and the prepare function will not be called
at all.

CRs-Fixed: 249142

Change-Id: I7eb1cd6ba7621f400ea2fc2da511d8c5afc927ae