platform/external/opencore: Improve Filewriter performance in opencore composer node
Fragment writer class uses the pmem buffer directly for file write. This adversely
affects performance at high bit rates as the encoder component is starved for output
buffers. This change implements a ring buffer that is used to copy the pmem data into
a local buffer. The local buffer is used for file write. The pmem buffer is released
back to the encoder as soon as possible.

Change-Id: If9e9f62a4bbfe30f9fd852639d422892332f302d