external/opencore: Fix for Rebuffering Issue
JitterBuffer would notify an underflow to PlayerEngine on
exhausting the buffers during a playback and if the playback
has not reached the EOS. PlayerEngine would pause the playback
on receiving Underflow notification. On rebuffering above the
buffering threshlod JitterBuffer would adjust the playback
time to new timestamp. JitterBuffer would issue a BOS to
audio and video decoders & set the playback clock.

Change-Id: Ifb9d951351d7c22bda768d9993743fc3aad17998
CRs-Fixed: 241893