external/opencore: Fix to call start cmnd after seek complete in case of resume
Afetr suspend,when the playback resumes,the data path is set up
in prepare and a seek to the position before suspend is done.
Currently, the playerdriver used to wait for a start cmnd from the
Mediaplayer.Change made to call start after seek complete is received
in playerdriver in case of resume.

Change-Id: I538e10f70e4d00875b38bc5e235fca11544856f0
CRs-Fixed: 256439