external/opencore: Fix loss of last few frames on playback of AMR content
AMR content is read in bundles of ten frames. So,if at the end of
the file,the remaining number of frames is less than ten,a
DATA_INSUFFICIENT reply is sent. This causes the parser to ignore these
last few frames. Instead if END_OF_FILE reply is sent,these frames are
are not discarded and are sent for decode.

Change-Id: Ida2c27e88aeabe8140ecfd2722e397cabbaac4f1
CRs-Fixed: 259944