external/opencore: Fix loss of last few frames on playback of AMR content
AMR content is read in bundles of ten frames. So,if at the end of
the file,the remaining number of frames is less than ten,a
DATA_INSUFFICIENT reply is sent. This causes the parser to ignore these
last few frames. Instead if END_OF_FILE reply is sent,these frames are
are not discarded and are sent for decode.

Change-Id: I54b389e716ed6966cbd4d595eb23630dae7d63bb
CRs-Fixed: 259944