external/opencore: Disable state change from PREPARE to PAUSE for streaming content
After PREPARE cmd complete is received,we put the player in
PAUSE state for audio only playback to handle TCXO shutdown.
This state transition is not required for streaming content as
playback starts immediately after preapre is complete.Added a
condition  to check for streaming content before state transition.

Change-Id: I83e2a29b73b3732b496812476adc815f07821095
CRs-Fixed: 261003