external/opencore: Disable state change from PREPARE to PAUSE for streaming content
After PREPARE cmd complete is received,we put the player in
PAUSE state for audio only playback to handle TCXO shutdown.
This state transition is not required for streaming content as
playback starts immediately after preapre is complete.Added a
condition  to check for streaming content before state transition.

Change-Id: Ibcfebf8ec5cad7c2a6a1332469b58cd85812b678
CRs-Fixed: 261003