external/opencore: Flush audio sink when the playback is paused
When playback is paused and a reposition is done, the playback begins
with few residual frames from the paused position and then the frames
from the new position. This is because the frames which were decoded
but not rendered when playback was paused are not flushed. Flushing
those frames by calling audio sink stop.

Change-Id: I0889138c6ce04f8afe3c79ea21946a19e4a7d87b
CRs-Fixed: 261610