external/opencore: Added check to avoid race condition when setting first timestamp
- CameraInput: Added a check in postWriteAsync() to verify that the first audio
timestamp is set not in this function, but only from ClockStateUpdated() when
recording with audio.
- AudioInput: Set the first frame received flag to true until after the first
timestamp is actually set.

This change prevents a race condition that caused the first audio timestamp obtained by
CameraMIO to become incorrect. This caused the recording process to be stopped immediately
upon receiving the first video frame, due to the wide time difference between the video
frame's timestamp and the incorrect first audio timestamp.

Change-Id: I983be6ed697333f14260bf2bf3d5339045e2e904
CRs-Fixed: 265103