external/opencore: Parse video encoder output frames only for SPS NAL Unit types
This change speeds up the parsing of H.264 output buffers by checking for NAL unit
types in Opencore's encoder node. If the type is SPS or PPS, it will continue looking
for the next NAL unit start code; if it is not one of these types, it stops parsing
for the next start code.

CRs-Fixed: 257540,267441

Change-Id: