external/opencore: Fix for clips with corrupted track atom size
While parsing a clip, if the track atom parse fails, it skips it
and continues with the next atom. But if the size of the failed
atom is zero, the parser cannot increment the filepointer and continuously
reads the same atom in an infinite loop. Adding a check for track atom size
and allow if its not zero or more than movie atom size.

Change-Id: I5d7b4a90774865d5cb67998d9866d2134f271c09
CRs-Fixed: 274113
(cherry picked from commit a18ab0a4764803d8851810a73b46d99af91d99e1)