external/opencore: Fix to resend the config buffer after reposition
If a seek is done immediately after start of playabck, the video track
config buffer that is sent to the OMX decoder may get flushed due to the
flush cmnd generated by the reposition. This scenario can occur when RTSP
streaming playback is suspended and resumed. Made changes to store the config
buffer sent to the video decoder and resend it each time after reposition.

CRs-Fixed: 270722

Change-Id: Ie3862fb85a6e7e8eff7349f1ace9085f0b01a9f3