external/opencore: Fix for corrupt nOffset value in output buffer
-A crash observed in Netflix streaming.
-When dynamic port reconfig happens, nOffset value
in o/p buffer is not set to zero by the aac omx component.
The o/p buffer is filled with the required o/p buffer
size as specified by the o/p port's nBufferSize and
given back to IL client without setting the nOffset
value to zero.
-Set the nOffset value to zero during dynamic port
reconfig.

Change-Id: I904f671b22eea086b51208169b172eafd0cda61c
CRs-Fixed: 309053