external/opencore: Fix for fragment clip playback failure when using URI
Calling setdatasource with URI on fragmented clip caused the clip to fail
after first few seconds of playback.

In opencore during seek while read, the offset is typecast to 64bit variable
type. When lseek is called with a negative offset, this translates to a huge
offset value thereby seek failing.

Change-Id: I5d5dc68dac8f5210914e3b8428cc25c9e3debf03
CRs-Fixed: 333628